Adding command line option '--reuse_containers'
This option needs a given 'name' container attribute which is used to search for an already existing container.
If such a container exists it will be reused. If not a new container will be created.
The finalizer for removing the container at the end of test execution is not added if --reuse_containers is set